GETP-99 test: 테스트 코드에 대해 변경 사항 반영

+++ src/test/java/es/princip/getp/fixture/PeopleFixture.java
@@ -60,7 +60,7 @@ public class PeopleFixture {
             .build()).toList();
     }
 
-    public static List<CardPeopleResponse> createCardPeopleResponses(Long count) {
+    public static List<CardPeopleResponse> createCardPeopleResponses(int count) {
         return LongStream.range(0, count)
             .mapToObj(i -> CardPeopleResponse.from(
                 i,
